% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetricsTripartiteSBM.R
\name{getMetricsTripartiteSBM}
\alias{getMetricsTripartiteSBM}
\title{Compute structural metrics on a tripartite network model MBM}
\usage{
getMetricsTripartiteSBM(myMSBM, normalise = TRUE)
}
\arguments{
\item{myMSBM}{the model to analyze}

\item{normalise}{enable metrics' normalization if TRUE}
}
\value{
a data.frame including for each cluster of each FG of the network the metrics values and the relative sizes of clusters \code{cluster_pi}.
}
\description{
Compute structural metrics on a tripartite network model MBM
}
\examples{
getMetricsTripartiteSBM(myMSBM)

}
