% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnalysisTripartiteSBM.R
\name{plotAnalysisTripartiteSBM}
\alias{plotAnalysisTripartiteSBM}
\title{Plot main blocks structural proprieties of a collection of tripartite networks,
that emerge from MBM parameters.}
\usage{
plotAnalysisTripartiteSBM(
  analyse,
  Indices = as.numeric(names(table(analyse$network_id))),
  new.window = TRUE
)
}
\arguments{
\item{analyse}{the output data.frame of our analysis,
including at least numbers of blocks \code{nb_herbi_clust}, \code{nb_plants_clust},\code{nb_polli_clust}
and relative sizes of blocks \code{cluster_pi}}

\item{Indices}{optional vector of networks' indexes to be plotted}

\item{new.window}{plots in a new window if TRUE}
}
\value{
a new figure
}
\description{
Plot main blocks structural proprieties of a collection of tripartite networks,
that emerge from MBM parameters.
}
\examples{
data(analyseEcologicalNetworksDataset)
plotAnalysisTripartiteSBM(analyse)

}
